Handled the following cases:
If a site goes down and the recovers, read on a non-replicated variable
can be returned. However, for a replicated variable, we need to wait
until commit happens.

In case every site failed between the last commit recorded on that site
and the time transaction began, then we need to abort the transaction.